# swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Contacts API",
      "description": "API for managing contacts",
      "version": "1.0.0"
    },
    "host": "https://cse341-node-hctg.onrender.com/",
    "basePath": "/",
    "schemes": [      
      "https"
    ],
    "paths": {
      "/contacts/": {
        "get": {
          "description": "Get all contacts",
          "responses": {
            "200": {
              "description": "OK"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        },
        "post": {
          "description": "Create a new contact",
          "parameters": [
            {
              "name": "contact",
              "in": "body",
              "description": "Contact data to be added",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "favoriteColor": {
                    "type": "string"
                  },
                  "birthday": {
                    "type": "string"
                  }
                },
                "required": ["firstName", "lastName", "email"]
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Created"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/contacts/{id}": {
        "get": {
          "description": "Get a contact by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "404": {
              "description": "Not Found"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        },
        "put": {
          "description": "Update a contact",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "type": "string"
            },
            {
              "name": "contact",
              "in": "body",
              "description": "Updated contact data",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "favoriteColor": {
                    "type": "string"
                  },
                  "birthday": {
                    "type": "string"
                  }
                },
                "required": ["firstName", "lastName", "email"]
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        },
        "patch": {
          "description": "Partially update a contact",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "type": "string"
            },
            {
              "name": "contact",
              "in": "body",
              "description": "Partial contact data to update",
              "required": false,
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "favoriteColor": {
                    "type": "string"
                  },
                  "birthday": {
                    "type": "string"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "404": {
              "description": "Not Found"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        },
        "delete": {
          "description": "Delete a contact by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      }
    }
  }
  